+++ addon/run.sh
@@ -15,6 +15,7 @@ export TIMER_MEDIA_PLAYER_ENTITY_ID=$(bashio::config 'timer_media_player_entity_
 export TIMER_DEFAULT_MEDIA_URL=$(bashio::config 'timer_default_media_url')
 export TIMER_DEFAULT_MEDIA_CONTENT_TYPE=$(bashio::config 'timer_default_media_content_type')
 export TIMER_DEFAULT_SCRIPT_ID=$(bashio::config 'timer_default_script_id')
+export TIMER_ALARM_REPEAT_INTERVAL_SECONDS=$(bashio::config 'timer_alarm_repeat_interval_seconds')
 export HA_URL="http://supervisor/core"
 export HA_TOKEN="${SUPERVISOR_TOKEN}"